+++ aquadoggo/src/proptests/strategies/filter.rs
@@ -7,8 +7,8 @@ use proptest::sample::select;
 use proptest::strategy::{BoxedStrategy, Just, Strategy};
 use proptest_derive::Arbitrary;
 
-use crate::proptests::schema_strategies::{SchemaField, SchemaFieldType};
-use crate::proptests::utils::{FieldName, HexString};
+use crate::proptests::strategies::document::HexString;
+use crate::proptests::strategies::schema::{FieldName, SchemaField, SchemaFieldType};
 
 /// Possible values used in filter arguments. `UniqueIdentifier` is a placeholder for values which
 /// can be derived at runtime in order to use identifiers which exist in on the node, these include